LEGEND: Implementing Logics access methods

+++ engines/legend/games/gateway/logic/logics.h
@@ -795,7 +795,7 @@ public:
 	/**
 	 * Return number of logics
 	 */
-	size_t size() const override {
+	int size() const override {
 		return LOGICS_COUNT;
 	}
 };
